# #21 Factorize word_vectors reading and preparation to speed up feature prepration