--- a/sherlock/features/word_embeddings.py
+++ b/sherlock/features/word_embeddings.py
@@ -6,21 +6,12 @@ from collections import OrderedDict
 
 # Input: a single column in the form of a pandas series
 # Output: ordered dictionary holding word embedding features
-def extract_word_embeddings_features(values):
+def extract_word_embeddings_features(values, word_to_embedding):
 
     num_embeddings = 50
     f = OrderedDict()
     embeddings = []
 
-    word_vectors_f = open('../sherlock/features/glove.6B.50d.txt', encoding='utf-8')
-    word_to_embedding = {}
-
-    for w in word_vectors_f:
-
-        term, vector = w.strip().split(' ', 1)
-        vector = np.array(vector.split(' '), dtype=float)
-        word_to_embedding[term] = vector
-    
     values = values.dropna()
 
     for v in values: