Simple move of code to factorize the file reading and processing
**Speed up feature preparation by 16**

Results before factorization:
```
Preparing feature extraction by downloading 2 files:
        
 ../sherlock/features/glove.6B.50d.txt and 
 ../sherlock/features/par_vec_trained_400.pkl.docvecs.vectors_docs.npy.
        
All files for extracting word and paragraph embeddings are present.
Extracting features for data column: 10
32460ms
Extracting features for data column: 20
36977ms
Extracting features for data column: 30
35734ms
```

Results after factorization:
```
Preparing feature extraction by downloading 2 files:
        
 ../sherlock/features/glove.6B.50d.txt and 
 ../sherlock/features/par_vec_trained_400.pkl.docvecs.vectors_docs.npy.
        
All files for extracting word and paragraph embeddings are present.
Extracting features for data column: 10
2029ms
Extracting features for data column: 20
2162ms
Extracting features for data column: 30
2149ms
```