@madelonhulsebos 2021-03-23
This is great @sylmorin-gbif. Thanks for contributing this significant runtime improvement!